+++ src/model/IngameObject.java
@@ -2,7 +2,7 @@ package model;
 
 import java.awt.Dimension;
 import java.awt.geom.Point2D;
-import java.awt.geom.Point2D.Float;
+import java.awt.geom.Point2D.Double;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -24,7 +24,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
     
     protected Boolean _isDestroyed = false;
     
-	protected Point2D.Float _position = null;
+	protected Point2D.Double _position = null;
 	protected Dimension _size = null;
 	protected Speed2D _speed = null;
 	protected ArrayList<CollisionBehaviour> _defaultColBehaviour = new ArrayList<>();
@@ -41,7 +41,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	 */
 	public IngameObject(GameField field) {
 		
-	    this(field, new Point2D.Float(0, 0), new Dimension(0, 0));
+	    this(field, new Point2D.Double(0, 0), new Dimension(0, 0));
 	}
 	
 	/**
@@ -50,7 +50,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	 * @param pos Позиция объекта.
 	 * @param dim Размеры объекта.
 	 */
-	public IngameObject(GameField field, Point2D.Float pos, Dimension dim) {
+	public IngameObject(GameField field, Point2D.Double pos, Dimension dim) {
 	    
 	    this(field, pos, dim, new Speed2D(0, 0));
 	}
@@ -62,7 +62,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	 * @param dim Размеры объекта.
 	 * @param speed Скорость объекта.
 	 */
-	public IngameObject(GameField field, Point2D.Float pos, Dimension dim, Speed2D speed) {
+	public IngameObject(GameField field, Point2D.Double pos, Dimension dim, Speed2D speed) {
 	    
 	    if (field == null || pos == null || dim == null || speed == null) {
 	        throw new NullPointerException();
@@ -70,7 +70,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	    
 	    this._field = field;
 	    this.setSize(dim);
-        this.setPosition(pos);
+            this.setPosition(pos);
 	    this.setSpeed(speed);
 	}
 	
@@ -108,16 +108,16 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	 * Получить позицию.
 	 * @return Текущая позиция.
 	 */
-	public Point2D.Float getPosition() {
+	public Point2D.Double getPosition() {
 
-		return (Point2D.Float) _position.clone();
+		return (Point2D.Double) _position.clone();
 	}
 	
 	/**
 	 * Задать позицию.
 	 * @param pos Новая позиция.
 	 */
-	public void setPosition(Point2D.Float pos) {
+	public void setPosition(Point2D.Double pos) {
 
 	    if (pos == null) {
 	        throw new NullPointerException();
@@ -153,9 +153,9 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	 * Сдвинуть объект.
 	 * @param delta Величина изменения позиции
 	 */
-	public void move(Point2D.Float delta) {
+	public void move(Point2D.Double delta) {
 		
-		this.setPosition(new Point2D.Float(this.getPosition().x + delta.x, this.getPosition().y + delta.y));
+		this.setPosition(new Point2D.Double(this.getPosition().x + delta.x, this.getPosition().y + delta.y));
 	}
 	
 	/**
@@ -341,7 +341,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 	//-------------------------------------------------------------------------------------------//
 	
 	@Override
-	public void positionChanged(Point2D.Float newpos) {
+	public void positionChanged(Point2D.Double newpos) {
 		
 	    this._position = newpos;
 	}
@@ -411,7 +411,7 @@ public abstract class IngameObject implements Cloneable, PositionChangeListener,
 		IngameObject clone = (IngameObject)super.clone();
 		clone._isDestroyed = this._isDestroyed;
 		clone._field = this._field;    // ссылка на поле просто копируется, да
-		clone._position = (Point2D.Float) this._position.clone();
+		clone._position = (Point2D.Double) this._position.clone();
 		clone._size = (Dimension) this._size.clone();
 		clone._speed = (Speed2D) this._speed.clone();
