+++ src/model/brick/BreakableBrick.java
@@ -2,7 +2,7 @@ package model.brick;
 
 import java.awt.Dimension;
 import java.awt.geom.Point2D;
-import java.awt.geom.Point2D.Float;
+import java.awt.geom.Point2D.Double;
 
 import model.GameField;
 import model.Speed2D;
@@ -17,18 +17,18 @@ public class BreakableBrick extends Brick {
 
 	public BreakableBrick(GameField field) {
 		
-	    this(field, new Point2D.Float(0, 0), new Dimension(0, 0));
+	    this(field, new Point2D.Double(0, 0), new Dimension(0, 0));
 	}
 	
 
-	public BreakableBrick(GameField field, Float pos, Dimension dim, Speed2D speed) {
+	public BreakableBrick(GameField field, Double pos, Dimension dim, Speed2D speed) {
         
 	    super(field, pos, dim, speed);
 	    this.addDefaultCollisionBehaviour(BehaviourDestroy.getInstance());
     }
 
 
-    public BreakableBrick(GameField field, Float pos, Dimension dim) {
+    public BreakableBrick(GameField field, Double pos, Dimension dim) {
         
         this(field, pos, dim, new Speed2D(0, 0));
     }
