+++ src/model/brick/Brick.java
@@ -1,7 +1,7 @@
 package model.brick;
 
 import java.awt.Dimension;
-import java.awt.geom.Point2D.Float;
+import java.awt.geom.Point2D.Double;
 
 import model.GameField;
 import model.IngameObject;
@@ -19,12 +19,12 @@ public abstract class Brick extends IngameObject {
 	    super(field);
 	}
 
-    public Brick(GameField field, Float pos, Dimension dim, Speed2D speed) {
+    public Brick(GameField field, Double pos, Dimension dim, Speed2D speed) {
         
         super(field, pos, dim, speed);
     }
 
-    public Brick(GameField field, Float pos, Dimension dim) {
+    public Brick(GameField field, Double pos, Dimension dim) {
         
         super(field, pos, dim);
     }
